+++ HISTORY.rst
@@ -9,6 +9,8 @@ picows Release History
 ------------------
 
 * Allow to override Host header through extra_headers in Core API and additional_headers in websockets API
+* #108: fix picows.websockets.InvalidStatus.response is the raw WSUpgradeResponse not websockets.Response
+* #108: process_request receives ServerHandshakeConnection object that doesn't have respond method.
 
 2.1.3 (2026-08-04)
 ------------------